cleanup code and stuff
