--- src/js/bootstrap.js
@@ -1,2 +1,0 @@
-import '../stylus/app.styl';
-import './components/_index';
