+++ src/js/builder.js
@@ -1,58 +1,72 @@
-import Vue from 'vue';
-import './bootstrap';
-import Builder from './builderManager';
-import { getSectionById, removeFromArray } from './util';
+import Section from './section';
+import installComponents from './components';
 
-Vue.use(Builder);
-const bus = new Vue();
-Vue.prototype.$bus = bus;
+const BUILDER_OPTIONS = {
+  docTitle: '',
+  sections: []
+};
 
-new Vue({
-  el: '#app',
-  data: {
-    docTitle: '',
-    editable: true,
-    elements: []
-  },
-  methods: {
-    module () {
-    }
-  },
-  created () {
-    this.$bus.$on('addElement', (el) => {
-      this.elements.push(el);
-    });
-    this.$bus.$on('editable', () => {
-      this.editable = !this.editable;
-      this.elements.forEach((el) => {
-        el.editable = this.editable;
-      })
-    });
-    this.$bus.$on('updateImage', (id, imgId, src) => {
-      const section = getSectionById(this.elements, id);
-      const imageId = Number(imgId);
-      section.images[imageId] = src;
-    });
-    this.$bus.$on('updateHref', (id, newHref) => {
-      const section = getSectionById(this.elements, id);
-      section.button.href = newHref;
-    });
-    this.$bus.$on('removeSection', (id) => {
-      const section = getSectionById(this.elements, id);
-      const index = this.elements.indexOf(section);
-      this.elements.splice(index, 1);
-    });
-    this.$bus.$on('addClass', (id, el, className) => {
-      const section = getSectionById(this.elements, id);
-      section[el].class.push(className);
-    });
-    this.$bus.$on('removeClass', (id, el, className) => {
-      const section = getSectionById(this.elements, id);
-      className.forEach((currentClass) => {
-        removeFromArray(section[el].class, currentClass);
-      });
-    });
-  },
-  mounted () {
+// To tell if it is installed or not
+let _Vue = null;
+let _builder = null;
+
+export default class Builder {
+  constructor (options) {
+    options = Object.assign({}, BUILDER_OPTIONS, options);
+    this.isEditing = true;
+    this.sections = options.sections || [];
   }
-});
+
+  add (options) {
+    this.sections.push(new Section(options));
+  }
+
+  find (id) {
+    return this.sections.find(s => s.id === id);
+  }
+
+  remove (id) {
+    const idx = this.sections.findIndex(s => s.id === id);
+    this.sections.splice(idx, 1);
+  }
+
+  static install (Vue, options = {}) {
+    // already installed
+    if (_Vue) {
+      return;
+    }
+
+    _builder = new Builder(options);
+    Vue.util.defineReactive(_builder, 'sections', _builder.sections);
+    Vue.util.defineReactive(_builder, 'isEditing', _builder.isEditing);
+
+    Vue.mixin({
+      provide: function providesBuilder () {
+        const provides = {};
+        if (this.$builder) {
+          provides.$builder = this.$builder;
+        }
+
+        if (this.$section) {
+          provides.$section = this.$section;
+        }
+
+        return provides;
+      },
+      beforeCreate () {
+        this.$builder = _builder;
+        if (this.$options.propsData && this.$options.propsData.id !== undefined) {
+          this.$section = this.$builder.find(this.$options.propsData.id);
+        }
+      },
+      mounted () {
+        Array.from(this.$el.querySelectorAll('.is-editable')).forEach((el) => {
+          el.contentEditable = 'true';
+        });
+      }
+    });
+
+    installComponents(Vue);
+    _Vue = Vue;
+  }
+};
