+++ src/js/index.js
@@ -1,0 +1,9 @@
+import Vue from 'vue';
+import '../stylus/app.styl';
+import Builder from './builder';
+
+Vue.use(Builder);
+
+new Vue({
+  el: '#app'
+});
